[ArrayManager] Implement .equals method (#39721)
